# cmd/open_im_api/docs/swagger.yaml
basePath: /
definitions:
...
  Open_IM_pkg_proto_sdk_ws.PublicUserInfo:
    properties:
      ex:
        type: string
      faceURL:
        type: string
      gender:
        type: integer
      nickname:
        type: string
      pubKey:
        type: string
      userID:
        type: string
    type: object
  Open_IM_pkg_proto_sdk_ws.UserInfo:
    properties:
      appMangerLevel:
        type: integer
      birth:
        type: integer
      createTime:
        type: integer
      email:
        type: string
      ex:
        type: string
      faceURL:
        type: string
      gender:
        type: integer
      globalRecvMsgOpt:
        type: integer
      nickname:
        type: string
      phoneNumber:
        type: string
      pubKey:
        type: string
      userID:
        type: string
    type: object
...
  base_info.GetSelfUserInfoReq:
    properties:
      operationID:
        type: string
      userID:
        type: string
    required:
    - operationID
    - userID
    type: object
  base_info.GetSelfUserInfoResp:
    properties:
      errCode:
        type: integer
      errMsg:
        type: string
    type: object
...
  base_info.GetUsersInfoReq:
    properties:
      operationID:
        type: string
      userIDList:
        items:
          type: string
        type: array
    required:
    - operationID
    - userIDList
    type: object
  base_info.GetUsersInfoResp:
    properties:
      errCode:
        type: integer
      errMsg:
        type: string
    type: object
  base_info.GetUsersOnlineStatusReq:
    properties:
      operationID:
        type: string
      userIDList:
        items:
          type: string
        maxItems: 200
        type: array
    required:
    - operationID
    - userIDList
    type: object
  base_info.GetUsersOnlineStatusResp:
    properties:
      data:
        items:
          $ref: '#/definitions/pbRelay.GetUsersOnlineStatusResp_SuccessResult'
        type: array
      errCode:
        type: integer
      errMsg:
        type: string
    type: object
...
  base_info.SetGlobalRecvMessageOptReq:
    properties:
      globalRecvMsgOpt:
        enum:
        - 0
        - 1
        - 2
        type: integer
      operationID:
        type: string
    required:
    - operationID
    type: object
  base_info.SetGlobalRecvMessageOptResp:
    properties:
      errCode:
        type: integer
      errMsg:
        type: string
    type: object
...
  base_info.Swagger400Resp:
    properties:
      errCode:
        example: 400
        type: integer
      errMsg:
        example: err msg
        type: string
    type: object
  base_info.Swagger500Resp:
    properties:
      errCode:
        example: 500
        type: integer
      errMsg:
        example: err msg
        type: string
    type: object
...
  base_info.UpdateSelfUserInfoReq:
    properties:
      birth:
        type: integer
      email:
        maxLength: 64
        type: string
      ex:
        maxLength: 1024
        type: string
      faceURL:
        maxLength: 1024
        type: string
      gender:
        enum:
        - 0
        - 1
        - 2
        type: integer
      nickname:
        maxLength: 64
        minLength: 1
        type: string
      operationID:
        type: string
      phoneNumber:
        maxLength: 32
        type: string
      publicKey:
        maxLength: 64
        type: string
      userID:
        maxLength: 64
        minLength: 1
        type: string
    required:
    - operationID
    - userID
    type: object
...
  base_info.UpdateUserInfoResp:
    properties:
      errCode:
        type: integer
      errMsg:
        type: string
    type: object
...
  pbRelay.GetUsersOnlineStatusResp_SuccessDetail:
    properties:
      platform:
        type: string
      status:
        type: string
    type: object
  pbRelay.GetUsersOnlineStatusResp_SuccessResult:
    properties:
      detailPlatformStatus:
        items:
          $ref: '#/definitions/pbRelay.GetUsersOnlineStatusResp_SuccessDetail'
        type: array
      status:
        type: string
      userID:
        type: string
    type: object
...
info:
  contact: {}
  description: open-IM-Server 的API服务器文档, 文档中所有请求都有一个operationID字段用于链路追踪
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: open-IM-Server API
  version: "1.0"
paths:
...
  /user/get_self_user_info:
    post:
      consumes:
      - application/json
      description: 传入ID获取自己的信息
      operationId: GetSelfUserInfo
      parameters:
      - description: im token
        in: header
        name: token
        required: true
        type: string
      - description: 请求体
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/base_info.GetSelfUserInfoReq'
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            allOf:
            - $ref: '#/definitions/base_info.GetSelfUserInfoResp'
            - properties:
                data:
                  $ref: '#/definitions/Open_IM_pkg_proto_sdk_ws.UserInfo'
              type: object
        "400":
          description: errCode为400 一般为参数输入错误, token未带上等
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "500":
          description: errCode为500 一般为服务器内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 获取自己的信息
      tags:
      - 用户相关
  /user/get_users_info:
    post:
      consumes:
      - application/json
      description: 根据用户列表批量获取用户信息
      operationId: GetUsersInfo
      parameters:
      - description: im token
        in: header
        name: token
        required: true
        type: string
      - description: 请求体
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/base_info.GetUsersInfoReq'
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            allOf:
            - $ref: '#/definitions/base_info.GetUsersInfoResp'
            - properties:
                Data:
                  items:
                    $ref: '#/definitions/Open_IM_pkg_proto_sdk_ws.PublicUserInfo'
                  type: array
              type: object
        "400":
          description: errCode为400 一般为参数输入错误, token未带上等
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "500":
          description: errCode为500 一般为服务器内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 获取用户信息
      tags:
      - 用户相关
  /user/get_users_online_status:
    post:
      consumes:
      - application/json
      description: 获取用户在线状态
      operationId: GetUsersOnlineStatus
      parameters:
      - description: im token
        in: header
        name: token
        required: true
        type: string
      - description: 请求体
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/base_info.GetUsersOnlineStatusReq'
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/base_info.GetUsersOnlineStatusResp'
        "400":
          description: errCode为400 一般为参数输入错误, token未带上等
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "500":
          description: errCode为500 一般为服务器内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 获取用户在线状态
      tags:
      - 用户相关
  /user/set_global_msg_recv_opt:
    post:
      consumes:
      - application/json
      description: 设置全局免打扰
      operationId: SetGlobalRecvMessageOpt
      parameters:
      - description: im token
        in: header
        name: token
        required: true
        type: string
      - description: globalRecvMsgOpt为全局免打扰设置0为关闭 1为开启
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/base_info.SetGlobalRecvMessageOptReq'
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/base_info.SetGlobalRecvMessageOptResp'
        "400":
          description: errCode为400 一般为参数输入错误, token未带上等
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "500":
          description: errCode为500 一般为服务器内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 设置全局免打扰
      tags:
      - 用户相关
  /user/update_user_info:
    post:
      consumes:
      - application/json
      description: 修改用户信息 userID faceURL等
      operationId: UpdateUserInfo
      parameters:
      - description: im token
        in: header
        name: token
        required: true
        type: string
      - description: 请求体
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/base_info.UpdateSelfUserInfoReq'
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/base_info.UpdateUserInfoResp'
        "400":
          description: errCode为400 一般为参数输入错误, token未带上等
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "500":
          description: errCode为500 一般为服务器内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 修改用户信息
      tags:
      - 用户相关
swagger: "2.0"